requirements.txt
python-dotenv==0.19.1
flask~=3.0.3
mongoengine==0.27.0
pymongo[srv]==4.0.1
flask_swagger_ui~=4.11.1
flask-cors==4.0.1
pydantic~=2.7.3
pyjwt==2.8.0
ollama==0.3.1
requests~=2.32.3
pandas==2.1.4
numpy>=1.26.0,<2.0.0
scikit-learn~=1.5.0